// service/src/common/enum/flow.enum.ts
export enum Flow {
  CHANGE_LANG = 'changeLang',
  PERSONAL_DETAILS = 'personalDetails',
  BE_ANONYMOUS = 'beAnonymous',
  KEEP_ANONYMOUS = 'keepAnonymous',
  KEEP_ANONYMOUS_FALLBACK = 'keepAnonymousFallback',
  DONT_REUSE_DETAILS = 'dontReuseDetails',
  DONT_REUSE_DETAILS_FALLBACK = 'dontReuseDetailsFallback',
  REUSE_DETAILS = 'reuseDetails',
  REUSE_DETAILS_FALLBACK = 'reuseDetailsFallback',
  SHARE_STORY = 'shareStory',
  SHARE_ADDITIONAL_INFO = 'shareAdditionalInfo',
  SHARE_ADDITIONAL_INFO_FALLBACK = 'shareAdditionalInfoallback',
  RESTART_FLOW = 'restartFlow',
  THANKS = 'thanks',
  SHARE_ANONYMOUS = 'shareAnonymous',
  LANG_NOT_VALID = 'langNotValid',
  USER_NAME = 'userName',
  USER_AGE = 'userAge',
  USER_GENDER = 'userGender',
  USER_DISABILITY = 'userDisability',
  SERVICE_UNAVAILABLE = 'serviceUnavailable',
  PUBLISH_CONSENT = 'publishConsent',
  PUBLISH_CONSENT_DENIED = 'publishConsentDenied',
  PUBLISH_CONSENT_GIVEN = 'publishConsentGiven',
  DEMOGRAPHICS = 'demographics',
  CONTACT_CONSENT_DENIED = 'contactConsentDenied',
  CONTACT_CONSENT_APPROVED = 'contactConsentApproved',
  CONTACT_CONSENT = 'contactConsent',
  GOODBYE = 'goodbye',
}